04/11 Updates
I spent 4 hours in class trying to figure out why python wasn't working. Hana realized it's probably because I manually deleted the code for responseBox submissions without accounting those changes to the rest of the code, especially the javascript. i went back on chatgpt and asked to edit the original html css and js to eliminate the responseBox and it worked!